backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==2.0.0
pydantic==2.5.0
aiosqlite==0.20.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1